+++ CHANGELOG.md
@@ -20,7 +20,8 @@ Not quite compatible with Octave yet. See #43 [octave support](https://github.co
 
 ### 0.9.25-dev
 
-
+* Support Octave identifiers in Octave mode. In Octave you can start
+  identifiers with an underscore, but you can't do it in MATLAB.
 
 ### 0.9.24